feat: refatorando actiongs para um arquivo proprio
