+++ ignite-reactjs-2023/03-ignite-timer/src/contexts/CyclesContext.tsx
@@ -1,5 +1,6 @@
 import { createContext, ReactNode, useState, useReducer } from "react";
-import { ActionTypes, Cycle, cyclesReducer } from "../reducers/cycles";
+import { Cycle, cyclesReducer } from "../reducers/cycles/reducer";
+import { ActionTypes, addNewCycleAction, interruptCurrentCycleAction, markCurrentCycleAsFinishedAction } from "../reducers/cycles/actions";
 
 interface CreateCycleData{
     task: string;
@@ -41,13 +42,7 @@ export function CyclesContextProvider({ children }: CyclesContextProviderProps){
     }
 
     function markCurrentCycleAsFinished(){
-
-        dispatch({
-            type: ActionTypes.MARK_CURRENT_CYCLE_AS_FINISHED,
-            payload: {
-                activeCycleId,
-            },
-        });
+        dispatch(markCurrentCycleAsFinishedAction());
     }
 
     function createNewCycle(data: CreateCycleData){
@@ -59,23 +54,12 @@ export function CyclesContextProvider({ children }: CyclesContextProviderProps){
             startDate: new Date(),
         };
 
-        dispatch({
-            type: ActionTypes.ADD_NEW_CYCLE,
-            payload: {
-                newCycle,
-            },
-        });
+        dispatch(addNewCycleAction(newCycle));
         setAmountSecondsPassed(0);
     }
 
     function interruptCurrentCycle(){
-        
-        dispatch({
-            type: ActionTypes.INTERRUPT_CURRENT_CYCLE,
-            payload: {
-                data: activeCycleId,
-            },
-        });
+        dispatch(interruptCurrentCycleAction());
     }
 
     return (
