+++ ignite-reactjs-2023/03-ignite-timer/src/reducers/cycles/reducer.ts
@@ -1,3 +1,5 @@
+import { ActionTypes } from "./actions";
+
 export interface Cycle {
     id: string;
     task: string;
@@ -10,12 +12,6 @@ export interface Cycle {
 interface CyclesState {
     cycles: Cycle[]
     activeCycleId: string | null
-}
-
-export enum ActionTypes {
-    ADD_NEW_CYCLE = "ADD_NEW_CYCLE",
-    INTERRUPT_CURRENT_CYCLE = "INTERRUPT_CURRENT_CYCLE",
-    MARK_CURRENT_CYCLE_AS_FINISHED = "MARK_CURRENT_CYCLE_AS_FINISHED"
 }
 
 export function cyclesReducer(state: CyclesState, action: any){
